reorganize search files
