+++ emphasis_education_app/src/components/Search/LiftedSearch.tsx
@@ -7,10 +7,10 @@ import { useLazyQuery } from '@apollo/react-hooks';
 import { Input, Icon } from 'react-native-elements';
 import styled from 'styled-components';
 
-import SearchResults from './SearchResults';
+import SearchResults from '../Presentational/SearchResults';
 
-import { ISearchUserPayload, ISearchInput } from '../types';
-import { SEARCH_USERS } from '../queries/SearchUsers';
+import { ISearchUserPayload, ISearchInput } from '../../types';
+import { SEARCH_USERS } from '../../queries/SearchUsers';
 
 const SearchContain = styled(View)`
   padding-top: 20px;
